Fefactor: 타입 및 index 정리
