+++ src/@types/binance.types.ts
@@ -1,0 +1,7 @@
+export interface IBinanceTicker {
+  e: string; // Event type
+  E: number; // Event time
+  s: string; // Symbol
+  c: string; //  Last price
+  Q: string; //  Last quantity
+}
