+++ src/@types/bithumb.types.ts
@@ -17,24 +17,6 @@ export interface IBithumbWsTicker {
   symbol: string;
 }
 
-export interface IBithumbTicker {
-  symbol: string;
-  opening_price: string;
-  closing_price: string;
-  min_price: string;
-  max_price: string;
-  units_traded: string;
-  acc_trade_value: string;
-  prev_closing_price: string;
-  units_traded_24H: string;
-  acc_trade_value_24H: string;
-  fluctate_24H: string;
-  fluctate_rate_24H: string;
-  binancePrice?: string;
-  date?: string;
-  time?: string;
-}
-
 export type IBithumbFetchCandlestick = [
   number,
   string,
