+++ src/components/ChartLeft/ChartInfo/index.ts
@@ -1,0 +1,1 @@
+export { ChartInfo } from './ChartInfo';
