+++ src/components/ChartLeft/TradingViewWidget/index.ts
@@ -1,0 +1,1 @@
+export { TradingViewWidget } from './TradingViewWidget';
