+++ src/components/Widget/ChartWidget/index.ts
@@ -1,0 +1,1 @@
+export { ChartWidget } from './ChartWidget';
