+++ src/components/Widget/TickerWidget/index.ts
@@ -1,0 +1,1 @@
+export { TickerWidget } from './TickerWidget';
