+++ src/components/common/Header/Header.tsx
@@ -3,8 +3,8 @@ import * as styled from './Header.styles';
 import getGlobalCoinData from 'api/getGlobalCoinData';
 import { useRecoilState } from 'recoil';
 import { globalCoinState } from 'recoil/atoms/globalCoinAtoms';
-import useFetchExchangeRate from 'hooks/binance/useFetchExchangeRate';
-import useTheme from 'hooks/useTheme';
+import { useFetchExchangeRate } from 'hooks/binance';
+import { useTheme } from 'hooks';
 import { HiSun, HiMoon } from 'react-icons/hi2';
 import { DiGithubAlt } from 'react-icons/di';
 import LogoDark from 'assets/images/Logo-Dark.svg';
