+++ src/hooks/binance/updateSocketDataWithBinance.ts
@@ -1,5 +1,5 @@
-import { IBinanceTicker } from './useBinanceTicker';
-import { ITicker } from '../../@types/common.types';
+import type { IBinanceTicker } from '../../@types/binance.types';
+import type { ITicker } from '../../@types/common.types';
 import { removeUSDT } from 'utils';
 
 export function updateSocketDataWithBinance(
