+++ src/hooks/useChartTickers.ts
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 import { Interval } from 'components/Widget/Widget.constants';
 import getChartWidgetData from 'api/getChartWidgetData';
 
-const useChartTickers = (pairId: string, interval: Interval) => {
+export const useChartTickers = (pairId: string, interval: Interval) => {
   return useQuery(
     [pairId, interval],
     () => getChartWidgetData(pairId, interval),
@@ -11,5 +11,3 @@ const useChartTickers = (pairId: string, interval: Interval) => {
     },
   );
 };
-
-export default useChartTickers;
