+++ src/hooks/useGetTicker.ts
@@ -1,10 +1,10 @@
 import { useRef, useState, useEffect } from 'react';
-import useBinanceTicker from 'hooks/binance/useBinanceTicker';
+import { useBinanceTicker } from 'hooks/binance';
 import { baseExchangeState } from 'recoil/atoms/commonAtoms';
 import { useRecoilValue } from 'recoil';
 import type { ITicker } from '../@types/common.types';
 import type { IBithumbWsTicker } from '../@types/bithumb.types';
-import { updateSocketDataWithBinance } from './binance/updateSocketDataWithBinance';
+import { updateSocketDataWithBinance } from './binance';
 import {
   changes,
   changesRatio,
@@ -16,29 +16,9 @@ import {
 import { fetchUpbitMarketCode } from 'api/upbit/fetchUpbitMarketCode';
 import { fetchBithumbTicker } from 'api/bithumb/fetchBithumbTicker';
 import { replaceDate } from 'utils';
+import { IUpbitMarketCode, IUpbitTicker } from '../@types/upbit.types';
 
-export interface IUpbitTicker {
-  code: string;
-  change: string;
-  trade_price: number;
-  opening_price: number;
-  signed_change_rate: number;
-  signed_change_price: number;
-  highest_52_week_price: number;
-  lowest_52_week_price: number;
-  acc_trade_price_24h: number;
-  acc_trade_volume_24h: number;
-  trade_date: string;
-  binancePrice?: string;
-}
-
-export interface IUpbitMarketCode {
-  market: string;
-  korean_name: string;
-  english_name: string;
-}
-
-export default function useGetTicker() {
+export function useGetTicker() {
   const baseExchange = useRecoilValue(baseExchangeState);
   const { binanceTickers } = useBinanceTicker();
   const [socketDatas, setSocketDatas] = useState<ITicker[]>([]);
